fix: Fix fail to receive file events on a new mount namespace
In the new version of the kernel module, the design of device number
filtering and user space assembly path has been adopted. The mount
namespace filtering in the old version of the kernel module is no
longer applicable. So the relevant code of mount namespace filtering is
deleted in the new version of the kernel module.

Log: Adapt to linglong app